feat(search): button on search results with filters redirection
